Made changes as requested.
